# Conflicts:
#	vlib/toml/checker/checker.v

+++ vlib/net/address.v
@@ -10,13 +10,13 @@ union AddrData {
 }
 
 const (
-	addr_ip6_any = [16]byte{init: byte(0)}
-	addr_ip_any  = [4]byte{init: byte(0)}
+	addr_ip6_any = [16]u8{init: u8(0)}
+	addr_ip_any  = [4]u8{init: u8(0)}
 )
 
-fn new_ip6(port u16, addr [16]byte) Addr {
+fn new_ip6(port u16, addr [16]u8) Addr {
 	a := Addr{
-		f: u16(AddrFamily.ip6)
+		f: u8(AddrFamily.ip6)
 		addr: AddrData{
 			Ip6: Ip6{
 				port: u16(C.htons(port))
@@ -29,9 +29,9 @@ fn new_ip6(port u16, addr [16]byte) Addr {
 	return a
 }
 
-fn new_ip(port u16, addr [4]byte) Addr {
+fn new_ip(port u16, addr [4]u8) Addr {
 	a := Addr{
-		f: u16(AddrFamily.ip)
+		f: u8(AddrFamily.ip)
 		addr: AddrData{
 			Ip: Ip{
 				port: u16(C.htons(port))
@@ -133,7 +133,7 @@ pub fn resolve_addrs(addr string, family AddrFamily, @type SocketType) ?[]Addr {
 
 			return [
 				Addr{
-					f: u16(AddrFamily.unix)
+					f: u8(AddrFamily.unix)
 					addr: AddrData{
 						Unix: resolved
 					}
